requirements.txt
langchain>=0.1.0,<0.2.0
langchain-google-genai>=0.0.7,<0.1.0
langchain-community>=0.0.28,<0.1.0
langchain-experimental>=0.0.55,<0.1.0
langchain-mongodb>=0.1.0,<0.2.0
google-generativeai>=0.3.2,<0.4.0
faiss-cpu>=1.7.4,<2.0.0
pandas>=2.1.4,<3.0.0
streamlit>=1.31.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
streamlit-folium>=0.15.1,<0.16.0
folium>=0.15.1,<0.16.0
Pillow>=10.2.0,<11.0.0
requests>=2.31.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
pymongo>=4.6.1,<5.0.0
